// src/generated/library/ReviewGame.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Account } from "../live/Account";
import type { GameSummary } from "../live/GameSummary";
import type { ReviewEvent } from "./ReviewEvent";

export type ReviewGame = { id: string, gameId: string, server: string, account: Account, game: GameSummary, queueName: string, timeline: Array<ReviewEvent> | null, source: string, timelineError: string | null, bookmarked: boolean, };